Split ddp grad test
